pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-course"
version = "0.2.0"
description = "Submission-ready OpenEnv environment for egocentric factory manipulation tasks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "numpy>=1.26.0",
    "openai>=1.40.0",
    "openenv-core>=0.2.3",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "egocentric_dataset_test",
    "egocentric_dataset_test.competition*",
    "server*",
]
